Clean up handling of request queue.
Requests are now only removed from the queue when the server sends a 
reply or event with a different sequence number (which we assume must be 
for a later event).  This prevents confusion when two requests are 
queued, the first gets a reply (and hence gets dequeued) and then an 
event comes in referencing that request (because the server hasn't yet 
processed the next one).

[originally from svn r9877]