Add support for the Generic Event Extension.
This modifies the core protocol by adding the possibility for events to
exceed 32 bytes in size.  I don't think I've ever seen such an event in
the wild, but one would cause havoc with xtruss's decoding routines, so
it seems prudent to support them just in case.  There's also a single
request: GEQueryVersion.