Remove need for xtruss to know number of events and errors in extensions.
Instead, it now assumes that each event is associated with the
extension whose first event is closest below (or equal to) the event
in question.  This was what it did for unknown events before, so this
change only impacts events that xtruss thinks it understands.

Specifically, before this change, if the server had two extensions, A
and B, and xtruss thought that extension A had more events than the X
server implemented (e.g. because the X server was older than xtruss),
then xtruss might misinterpret some events for extension B as
belonging to extension A instead.  Now it will do the right thing, at
least as long as it's seen all the relevant QueryExtension replies.

The above two paragraphs also apply to errors, with appropriate
substitutions.