Fix a NULL-dereference in -p mode.
In -p mode, we trace an X client which we haven't watched from the
start, so we'll sometimes see responses or errors relating to requests
we never saw. This causes xlog_respond_to() to be called with
req==NULL; it does check for that possibility, but not until after
it's already checked req->printed. Oops.