Remove the master GNUmakefile.
More or less copied from the way I did it in agedu in 2011 but never
got round to here. The autotools makefile is now the One True Way; it
will autodetect the presence of Halibut and use it to decide whether
the man page is a rebuildable target or not.

(So if you start from the distribution tarball which contains the man
page already built, then it doesn't matter if you have Halibut
installed; but if you check out from git, you do need Halibut, because
the man page isn't there already.)