Tighten up parsing of -p operand.
Parsing a string into a number with sscanf will succeed if the string
has a _prefix_ that can be interpreted in the specified way, but it
will say nothing about whether that prefix was followed by anything
else.

I was trying to introduce a mode called '-p all' to trace all X
clients at once via X RECORD, and found that '-p all' is already
parsed without error as 0xA, with the 'll' on the end quietly ignored.

Now I use a %n format to check we ended up scanning the whole string,
so that only valid hex numbers will be interpreted as hex.