Add '-p all' and a couple of its friends.
The X RECORD extension has three special modes in addition to
recording a specific X client's traffic. One means 'record all X
clients'; one means 'record all clients connected at the time of
starting to record, but not any that connect after that'; and one is
the opposite of that, meaning 'record all X clients that connect after
this point, but ignore the ones already connected at the time of
starting to record'.

Those all seem like useful modes in xtruss too, so I've implemented
them as '-p all', '-p current' and '-p future' respectively.

The way you signal these special modes in the RECORD protocol is by
passing the special values 3, 1 and 2 (respectively) in the slot for
the client id to watch. So you could already have got xtruss to
implement the basic functionality of these new options via '-p 3', '-p
1' and '-p 2' respectively, by knowing what the protocol-specified
special values are. But the special names also trigger other behaviour
changes, like not exiting as soon as a 'client disconnect' signal is
seen (we want to keep recording all the _other_ clients!), and
printing the client id on every trace line right from the start.

(Not to mention, of course, that '-p all' is easier to remember!)