-p mode: tolerate 'client has connected' record type.
Once we start recording multiple X clients at once, the server will
send us records that indicate a new client connecting as well as an
existing one going away. We don't need to _do_ anything with them, but
we do need not to interpret them as protocol confusion and die.