-p mode: prepare to trace multiple client ids.
I'm gearing up to trace multiple clients at once using X RECORD. So
each client will need a separate 'struct xlog' tracking its state, and
we have to hold those in a tree234 sorted by their client id. Hence,
when we receive a data record from the server, we need to look at the
client id in that record and use it to look up the right xlog in our
tree.