Update and restructure to the latest PuTTY code.
xtruss has always been a spin-off project of PuTTY. It was derived,
once, from an old version of the PuTTY code, namely git commit
d58d1692b8fca54fa2a2bebc60ebcda386b71895. That code is now very out of
date, and I've been thinking for ages that it would be good to update
it, especially after all the major refactoring that happened in PuTTY
in the last couple of years has actually made it _more_ convenient to
reuse the library facilities in spinoff projects. Not only that, but
some directly useful functionality has been added in PuTTY, such as
setting up a local X11 server port to listen on, and writing out its X
authority file, so I can throw away xtruss's separate version of that.

This commit brings things up to date: xtruss is now based on PuTTY git
commit 7003b43963aef6cdf2841c2a882a684025f1d806, which is the current
state of development as of now. Moreover, the source files shared with
the PuTTY code base are all identical to their states upstream, except
for a very small number of removed sections clearly marked with

    #if 0 /* removed in xtruss */
    ...
    #endif /* removed in xtruss */

which should make it easy to bring in updates in future.

There was essentially no way to do this by means of a proper git merge
- not only has too much changed in the meantime, but xtruss's
directory layout was simplified from PuTTY's so even the most obvious
parts of the merge didn't work usefully. So instead I've lifted the
xtruss layer off the top of the old code, and transplanted it on to
the new code, essentially rewriting the glue layers that interface to
PuTTY's internal abstractions (Plug and the like), but leaving the
main X tracing logic unchanged.

In the course of that, I've also taken the opportunity to break up the
xtruss source itself into multiple source files. Most of them - all
the interesting stuff! - now live in the platform-independent
top-level directory, while the official main() and the very small
number of Unix-specific parts live in unix/uxxtruss.c.

This fulfills the TODO item from the big comment in the old monolithic
uxxtruss.c that it would be nice to separate out the platform-
dependent parts. This opens the way to write a Windows port of xtruss,
should anyone ever need it!