Clean up handling of the signal pipe.
I wrote into it in the signal handler, but forgot to read the bytes
back out in the poll loop! Also the pipe wasn't set nonblock (without
which _both_ those operations are fraught with danger).

While I'm here, enlarged on the comment saying why the error check in
the signal handler does nothing.