-p mode: load existing atoms from X server on startup
This should get most atom references logged using their names, not IDs.
Some corner cases, when unresolved IDs could be seen, are still possible
but they should be quite rare.
In my local testing the loading happened in an instant for a thousand atoms.
A command-line option to disable the loading is added anyway, just in case
it'll take more time in some other setup or cause any problems.