Fix handling of EADDRINUSE.
In the loop searching for a free TCP port to run an X display on, we
should have continued to the next possible value if we _did_ get
EADDRINUSE, and broken the loop if we failed to bind to a port for any
other reason. But I got the sense of a strcmp the wrong way round, so
we did the opposite. Oops.