Update to current PuTTY main branch.
This brings us up to commit bff0c590e5e857d3ce06d454f260d4263e10cc05
in PuTTY, which was the tip of main as of starting this update.

This merge moves past a huge reorganisation of the PuTTY source code,
which broke up many previously monolithic source modules, moved lots
of files into subdirectories, renamed many files to have more sensible
names, and most importantly, migrated the build system to CMake, which
involved in particular collecting many object files into libraries so
that the makefiles no longer have to specify the exact set of objects
used in each program.

For the sake of keeping consistency with the upstream project, I've
therefore migrated xtruss to cmake as well, using something as similar
as possible to PuTTY's cmake system (but cut down considerably for
simplicity). The alternative would have been to either add library
support to the autotools makefile (making it more complicated in
itself), or else to figure out the precise set of object files needed
by xtruss and write them down in Makefile.am (introducing the
maintenance burden of re-checking all that on future merges). I think
moving to CMake is the easiest approach all round, and keeps things
similar enough that further merges should cause the least trouble out
of those options.

One side effect of the improved modularity of the upstream code is
that I've been able to throw out a few more pieces of PuTTY that
weren't _really_ used in xtruss but had to be included to prevent link
failures. In particular, the entire cryptographic random number
generator is gone (I think the code that previously needed it was
functions in x11fwd.c that were never actually called by xtruss), and
so has be_misc.c.

On the other hand, I haven't done a very thorough cleanup of the utils
subdirectory. I've got rid of things that really are specific to
things PuTTY does and xtruss doesn't (like SSH key fingerprints, or
login prompts, or base64), but a fair amount of general-purpose
library code is still in there and currently unused (like chomp(), or
nullstrcmp()). This doesn't add a maintenance burden because it's
identical to the upstream code, and it means that if I find a use for
those functions later, I won't have to faff about with bringing them
in from PuTTY in a special commit.